refactoring: namespace IGameReviewService, GameReviewService
