+++ GameStore.BLL/Services/GamesServices/GameReviewService.cs
@@ -8,7 +8,7 @@ using GameStore.DAL.Entities.Games;
 using GameStore.BLL.DTO.Games;
 using GameStore.DAL.Entities.History;
 
-namespace GameStore.BLL.Services.GameReviewServices
+namespace GameStore.BLL.Services.GamesServices
 {
     public class GameReviewService : IGameReviewService
     {
@@ -26,7 +26,7 @@ namespace GameStore.BLL.Services.GameReviewServices
             ResultServiceModel result = new();
 
             AppUser user = await _dbContext.AppUsers.Where(x => x.Email == model.UserEmail).FirstOrDefaultAsync();
-            if(user is null) { result.IsSucceeded = false; return result; }
+            if (user is null) { result.IsSucceeded = false; return result; }
 
             GameReview gameReview = new()
             {
@@ -48,22 +48,22 @@ namespace GameStore.BLL.Services.GameReviewServices
             return result;
         }
 
-        public async Task<List<GameReviewDTO>> GetGameReviewsAsync(long gameId, int pageNumber, int pageSize) 
+        public async Task<List<GameReviewDTO>> GetGameReviewsAsync(long gameId, int pageNumber, int pageSize)
         {
             List<GameReview> reviews = await _dbContext.GameReviews.Where(x => x.GameId == gameId && x.isDelete == false).OrderByDescending(x => x.DateCreate)
                                                                    .Include(u => u.User)
                                                                    .Skip((pageNumber - 1) * pageSize)
                                                                    .Take(pageSize).ToListAsync();
-          
+
             return _mapper.Map<List<GameReviewDTO>>(reviews);
         }
 
-        public async Task<ResultServiceModel> ElasticRemoveReviewAsync(long reviewId) 
+        public async Task<ResultServiceModel> ElasticRemoveReviewAsync(long reviewId)
         {
             ResultServiceModel result = new();
 
             GameReview review = await _dbContext.GameReviews.Where(x => x.Id == reviewId).FirstOrDefaultAsync();
-            if(review is null) { result.IsSucceeded = false; return result; }
+            if (review is null) { result.IsSucceeded = false; return result; }
 
             review.isDelete = true;
 
@@ -88,7 +88,7 @@ namespace GameStore.BLL.Services.GameReviewServices
             ResultServiceModel result = new();
 
             GameReview review = await _dbContext.GameReviews.Where(x => x.Id == gameReview.Id).FirstOrDefaultAsync();
-            if(review is null) { result.IsSucceeded = false; result.ErrorMes = DefaultErrorMessages.recordNoExist; return result; }
+            if (review is null) { result.IsSucceeded = false; result.ErrorMes = DefaultErrorMessages.recordNoExist; return result; }
 
             //История
             HistoryGameReviews history = new();
@@ -108,8 +108,8 @@ namespace GameStore.BLL.Services.GameReviewServices
                 await _dbContext.SaveChangesAsync();
             }
             catch { result.IsSucceeded = false; result.ErrorMes = DefaultErrorMessages.dontSave; return result; }
-         
-            
+
+
 
             result.IsSucceeded = true;
             return result;
