+++ GameStore.WEB/Controllers/HomeController.cs
@@ -3,9 +3,8 @@ using GameStore.BLL.DTO.Games;
 using GameStore.BLL.Infrastrcture;
 using GameStore.BLL.Infrastrcture.Models;
 using GameStore.BLL.Predefined;
-using GameStore.BLL.Services.GameReviewServices;
+using GameStore.BLL.Services.GamesServices;
 using GameStore.BLL.Services.HomeServices;
-using GameStore.DAL.Entities.Games;
 using GameStore.WEB.Infrastrcture;
 using GameStore.WEB.Models.HomeModels.HomePageModels;
 using Microsoft.AspNetCore.Authorization;
