+++ GameStore.WEB/Program.cs
@@ -24,7 +24,6 @@ using GameStore.BLL.Services.HomeServices;
 using GameStore.BLL.Services.ShoppingCartServices;
 using GameStore.BLL.Services.CatalogServices;
 using GameStore.BLL.Infrastrcture.AutomapperProfiles.HistoryProfiles;
-using GameStore.BLL.Services.GameReviewServices;
 
 var builder = WebApplication.CreateBuilder(args);
